refactor: move sink to util
